Wa 344: Fetch Mp Id from F5 page to place bets (#9)
* WA-344 : Fetch Mp Id from F5 page to place bets
Please note this code is part of a step, not a step itself. Mp id will be used internally for placing bet. We wont need to mention it in feature file.

* Wa 344: Fetch Mp Id from F5 page to place bets
Please note this code is part of a step, not a step itself. Mp id will be used internally for placing bet. We wont need to mention it in feature file.

* Wa 344: Changes mentioned in review and few more.

* Wa 344: Renamed webelement.

* Wa 344: Commit is failing. Hence adding another change, just to commit again.